swagger, controllers

+++ src/main/java/ru/sber/internship/service/impl/ClientServiceImpl.java
@@ -21,17 +21,32 @@ public class ClientServiceImpl implements ClientService {
     }
 
     @Override
-    public Client findById(int id) {
+    public Client findById(long id) {
         return clientRepository.findById(id);
     }
 
-    @Override
-    public void deleteById(int id) {
+
+    public boolean deleteById(long id) {
+        if (clientRepository.findById(id) == null) {
+            return false;
+        }
         clientRepository.deleteById(id);
+        return true;
     }
 
-
-    public void save(Client client) {
-        clientRepository.save(client);
+    public Client save(Client client) {
+        return clientRepository.save(client);
     }
+
+//    public Client getClientFromRequest(JsonNode node, ObjectMapper mapper) {
+//        JsonNode clientNode = node.get("client");
+//        Client client = null;
+//        try {
+//            client = mapper.treeToValue(clientNode, Client.class);
+//        } catch (JsonProcessingException e) {
+//            e.printStackTrace();
+//        }
+//
+//        return client;
+//    }
 }
