+++ src/main/java/ru/sber/internship/service/impl/OrderServiceImpl.java
@@ -4,11 +4,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.sber.internship.entity.Client;
 import ru.sber.internship.entity.Order;
-import ru.sber.internship.entity.Product;
 import ru.sber.internship.repository.OrderRepository;
-import ru.sber.internship.repository.ProductRepository;
 import ru.sber.internship.service.OrderService;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -23,17 +22,29 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public Order findById(int id) {
+    public Order findById(long id) {
         return orderRepository.findById(id);
     }
 
-    public void save(Order order) {
-        orderRepository.save(order);
+
+    public Order save(Order order) {
+        return orderRepository.save(order);
     }
 
     @Override
-    public void deleteById(int id) {
+    public BigDecimal calcTotalPrice(Order order) {
+       return order.getOrderItems().stream().
+                map(p -> p.getProduct().getPrice().multiply(new BigDecimal(p.getCount()))).
+                reduce(BigDecimal.ZERO, BigDecimal::add);
+
+    }
+
+    public boolean deleteById(long id){
+        if (orderRepository.findById(id) == null){
+            return false;
+        }
         orderRepository.deleteById(id);
+        return true;
     }
 
     @Override
@@ -42,4 +53,15 @@ public class OrderServiceImpl implements OrderService {
     }
 
 
+//    public Order getOrderFromRequest(JsonNode node, ObjectMapper mapper) {
+//        JsonNode orderNode = node.get("order");
+//        Order order = null;
+//        try {
+//            order = mapper.treeToValue(orderNode, Order.class);
+//        } catch (JsonProcessingException e) {
+//            e.printStackTrace();
+//        }
+//
+//        return order;
+//    }
 }
